// web/src/gen/types/CompleteRegister.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

/**
 * @description Default Response
 */
export type CompleteRegister204 = any;

/**
 * @description Default Response
 */
export type CompleteRegister406 = {
  /**
   * @type string
   */
  message: string;
};

/**
 * @description Default Response
 */
export type CompleteRegister409 = {
  /**
   * @type string
   */
  message: string;
};

/**
 * @description Default Response
 */
export type CompleteRegister500 = {
  /**
   * @type string
   */
  message: string;
};

export type CompleteRegisterMutationRequest = {
  /**
   * @minLength 10
   * @maxLength 11
   * @type string
   */
  phone: string;
  /**
   * @minLength 11
   * @maxLength 11
   * @type string
   */
  cpf: string;
  /**
   * @minLength 14
   * @maxLength 18
   * @type string | undefined
   */
  cnpj?: string;
  /**
   * @type string
   */
  street: string;
  /**
   * @type string
   */
  number: string;
  /**
   * @type string
   */
  neighborhood: string;
  /**
   * @type string
   */
  city: string;
};

export type CompleteRegisterMutationResponse = CompleteRegister204;

export type CompleteRegisterMutation = {
  Response: CompleteRegister204;
  Request: CompleteRegisterMutationRequest;
  Errors: CompleteRegister406 | CompleteRegister409 | CompleteRegister500;
};